+++ test/qnitensor.jl
@@ -1,4 +1,5 @@
 using ITensors
+using ITensors.NDTensors
 using LinearAlgebra
 using Random
 using Test
@@ -1504,7 +1505,7 @@ Random.seed!(1234)
       for b in nzblocks(V)
         @test flux(V, b) == QN()
       end
-      @test U * S * V ≈ A atol = 1e-15
+      @test U * S * V ≈ A atol = 1e-13
     end
 
     @testset "SVD no truncate bug" begin
@@ -1684,6 +1685,29 @@ Random.seed!(1234)
         @test expA[n, n] == exp(A[n, n])
       end
       @test expA ≈ exp(dense(A))
+    end
+
+    @testset "diag" for ElType in (Float64, ComplexF64)
+      χ = [QN(0) => 1, QN(1) => 2]
+      i, j = Index.((χ,), ("i", "j"))
+      A = randomITensor(ElType, i, j)
+      d = diag(A)
+      @test d isa DenseTensor{ElType,1}
+      for n in 1:dim(χ)
+        @test d[n] == A[n, n]
+      end
+    end
+
+    @testset "diag" for ElType in (Float64, ComplexF64)
+      χ = [QN(0) => 1, QN(1) => 2]
+      i, j = Index.((χ,), ("i", "j"))
+      A = randomITensor(ElType, i, j)
+      _, S, _ = svd(A, i)
+      d = diag(S)
+      @test d isa DenseTensor{real(ElType),1}
+      for n in 1:diaglength(S)
+        @test d[n] == S[n, n]
+      end
     end
 
     @testset "Mixed arrows" begin